backend/src/domain/use-cases/get-user-by-id-usecase.ts
import { User } from '../entities/user'
import { UsersRepository } from '../repositories/users-repository'
import { UserNotFoundError } from './errors/user-not-found-error'

interface GetUserByIdUseCaseRequest {
  userId: string
}

interface GetUserByIdUseCaseResponse {
  user: User
}

export class GetUserByIdUseCase {
  constructor(private usersRepository: UsersRepository) {}

  async execute({
    userId,
  }: GetUserByIdUseCaseRequest): Promise<GetUserByIdUseCaseResponse> {
    const user = await this.usersRepository.findById(userId)

    if (!user) {
      throw new UserNotFoundError()
    }

    return {
      user,
    }
  }
}
